requirements.txt
flask
flask-cors
mysql-connector-python
opencv-python
dlib
face_recognition
pandas
openpyxl
python-dotenv
